backend_services/backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary
python-dotenv==1.0.1
pydantic
httpx==0.27.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]
bcrypt==4.0.1